moved getFiles into endpoints/getFiles package
